+++ tenmo-server/src/main/java/com/techelevator/tenmo/model/TransferStatus.java
@@ -1,19 +1,19 @@
 package com.techelevator.tenmo.model;
 
 public class TransferStatus {
-    private int transferStatusId;
+    private long transferStatusId;
     private String transferStatusDesc;
 
-    public TransferStatus(int transferStatusId, String transferStatusDesc) {
+    public TransferStatus(long transferStatusId, String transferStatusDesc) {
         this.transferStatusId = transferStatusId;
         this.transferStatusDesc = transferStatusDesc;
     }
 
-    public int getTransferStatusId() {
+    public long getTransferStatusId() {
         return transferStatusId;
     }
 
-    public void setTransferStatusId(int transferStatusId) {
+    public void setTransferStatusId(long transferStatusId) {
         this.transferStatusId = transferStatusId;
     }
